% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_funs.R
\name{cmp_funs}
\alias{cmp_funs}
\alias{dist_euclidean}
\alias{dist_manhattan}
\alias{dist_chebyshev}
\alias{dist_bhat}
\alias{cor_pearson}
\alias{diff_rmse}
\alias{diff_mean}
\alias{diff_median}
\alias{diff_var}
\alias{diff_cv}
\alias{p_student}
\alias{p_wilcoxon}
\alias{dist_euclidean_cppr}
\alias{dist_manhattan_cppr}
\alias{dist_chebyshev_cppr}
\alias{diff_rmse_cppr}
\title{Summary metrics on a pair of rasters}
\usage{
dist_euclidean(x, y, ...)

dist_manhattan(x, y, ...)

dist_chebyshev(x, y, ...)

dist_bhat(x, y, bins = 10, ...)

cor_pearson(x, y, ...)

diff_rmse(x, y, ...)

diff_mean(x, y, ...)

diff_median(x, y, ...)

diff_var(x, y, ...)

diff_cv(x, y, ...)

p_student(x, y, min_nb = 5, ...)

p_wilcoxon(x, y, min_nb = 5, ...)

dist_euclidean_cppr(x, y, ...)

dist_manhattan_cppr(x, y, ...)

dist_chebyshev_cppr(x, y, ...)

diff_rmse_cppr(x, y, ...)
}
\arguments{
\item{x}{a fraction of a \code{SpatRaster} raster}

\item{y}{a fraction of a \code{SpatRaster} raster}

\item{...}{additional arguments}

\item{bins}{number of bins to use}

\item{min_nb}{minimal number of observation to calculate a value,
otherwise a \code{NA_real} is returned}
}
\value{
numeric the index of interest
}
\description{
Summary functions that return a scalar when passed with two portions of two rasters.
Typically passed to \link{CMP}.
}
\details{
First of all, you can use built-in function,
as long as they return a single value such as \code{cor}.
\itemize{
\item \code{cor_pearson} is pearson correlation coefficient. Prone to NA for small window sizes with sd=0.
\item \code{dist_euclidean/dist_euclidean_cppr} euclidean distance (sqrt of sums of squared diff pixel-wise, divided by the number of valid pixels)
\item \code{dist_manhattan/dist_manhattan_cppr} manhattan distance (sum of absolute distances pixel-wise, , divided by the number of valid pixels)
\item \code{dist_chebyshev/dist_chebyshev_cppr} Chebyshev's distance (max of absolute distances pixel-wise)
\item \code{dist_bhat} Bhattacharyya's distance based on distribution distances calculated using nbins
\item \code{diff_rmse/diff_rmse_cppr} root-mean square error pixel-wise
\item \code{diff_mean} difference of mean values
\item \code{diff_mean} difference of mean values
\item \code{diff_var} difference of variance values
\item \code{diff_cv} difference of coefficient of variation values
\item \code{p_student} p value from a Student t test
\item \code{p_wilcoxon} p value from a Wilcoxon rank sum test
}

Naturally, you can also use built-in function, as long as they return a single value such as: \code{cor},
}
\examples{
set.seed(2329) # for reproducibility
x <- sample(1:3, size=49, replace=TRUE, prob=c(0.1, 0.1, 1))  # unbalanced vector
y <- sample(1:3, size=49, replace=TRUE, prob=c(0.1, 0.1, 1))  # another one

# examples runs mostly to test and exemplify
# but in CMP they will run on each pair of focal windows

# built in functions
cor(x, y)
cov(x, y)

# distance indices
dist_euclidean(x, y)
dist_euclidean_cppr(x, y)

dist_manhattan(x, y)
dist_manhattan_cppr(x, y)

dist_chebyshev(x, y)
dist_chebyshev_cppr(x, y)

dist_bhat(x, y)

# difference indices
diff_rmse(x, y)
diff_rmse_cppr(x, y)

diff_mean(x, y)
diff_median(x, y)
diff_var(x, y)
diff_cv(x, y)

# p value from Student t test
p_student(x, y)
p_wilcoxon(x, y)

}
\seealso{
\link{CMP}, \link{mhm_funs}
}
