% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_raster}
\alias{export_raster}
\title{Export raster to image format}
\usage{
export_raster(x, filename, ...)
}
\arguments{
\item{x}{\code{raster} or \code{SpatRaster}}

\item{filename}{where to write the file}

\item{...}{additional arguments to pass to \link[terra:writeRaster]{terra::writeRaster}}
}
\value{
writes a file
}
\description{
Exports a \code{raster} or \code{SpatRaster} to \code{.tif} or \code{.txt}
to stick to original file formats of legacy MHM/CMP softwares.
}
\examples{
# we don't run the export here to not mess CRAN/github
# if you do, this will write a file on your machine
\dontrun{
landscape <- import_example("l1.tif")
landscape \%>\% export_raster("landscape.tif")
}
}
