% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_txt}
\alias{export_txt}
\alias{export_txt_grid}
\title{Export raster to txt format}
\usage{
export_txt(x, filename, ...)

export_txt_grid(x, filename, signif = NULL, ...)
}
\arguments{
\item{x}{\code{raster} or \code{SpatRaster}}

\item{filename}{where to write the file}

\item{...}{additional arguments to pass to \link[utils:write.table]{utils::write.table}}

\item{signif}{how many significant numbers to retain (default to \code{NULL}, no rounding)}
}
\value{
writes a file
}
\description{
Exports a \code{raster} or \code{SpatRaster} to text format such as \code{.txt}
to stick to original file formats of legacy MHM/CMP softwares.
}
\examples{
# we don't run the export here to not mess CRAN/github
# if you do, this will write a file on your machine
\dontrun{
landscape <- import_example("l1.tif")
landscape \%>\% export_txt("landscape.txt")
}

}
