% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gmean}
\alias{gmean}
\title{Geometric mean}
\usage{
gmean(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
Brings a geometric mean function (\verb{prod(x)^(1/length(x)})
}
\details{
Geometric mean is possibly a better alternative than a bare \code{prod} when calculating
composite maps from monoscale maps.
}
\examples{
x <- c(1, 2, 3, 4, 6)
mean(x)  # arithmetic mean
gmean(x) # geometric mean
}
