% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_example}
\alias{import_example}
\title{Import example files}
\usage{
import_example(name)
}
\arguments{
\item{name}{of the example file. If empty a list of available examples is returned.}
}
\value{
\code{SpatRaster} ready to use with \link{MHM}/\link{CMP}
}
\description{
Imports one of the example files bundled with CMP
}
\details{
Thin wrappers around \code{system.file("extdata", "name_of_file", package="mucha")}.
Intended to ease working on example files and make them less verbose in the examples/vignette.
}
\examples{

# when called with no name, returns a list of available examples
import_example()

# then you can
landscape <- import_example("l1.tif")
p(landscape)
}
