% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_raster}
\alias{import_raster}
\title{Import image files to raster}
\usage{
import_raster(x)
}
\arguments{
\item{x}{path to the raster in a format accepted by \link[terra:rast]{terra::rast} will work.}
}
\value{
\code{SpatRaster} ready to use with \link{MHM}/\link{CMP}
}
\description{
Imports raster files using a thin wrapper around \link[terra:rast]{terra::rast}
}
\details{
An arbitrary crs is explicitely declared and warnings
about missing extent are supressed but declared anyway.
}
\examples{
# use a system.file to make it work here,
# otherwise just point to your file
# these two are smaller (resampled 0.5) versions of their tif counterpart
# because of data volume limitations

system.file("extdata/l1.tif", package = "mucha") \%>\%
import_raster() \%>\% p()

system.file("extdata/l2.tif", package = "mucha") \%>\%
import_raster() \%>\% p()

}
