% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_txt}
\alias{import_txt}
\alias{import_txt_grid}
\title{Import text file into raster}
\usage{
import_txt(x, header = FALSE, na.strings = "NA", sep = " ", ...)

import_txt_grid(x)
}
\arguments{
\item{x}{character path to file}

\item{header}{logical (default to \code{FALSE}) passed to \code{\link[=read.table]{read.table()}}}

\item{na.strings}{character (default to \code{-9999}) passed to \code{\link[=read.table]{read.table()}}}

\item{sep}{character (default to \code{" "}) passed to \code{\link[=read.table]{read.table()}}}

\item{...}{more params to be passed to \code{\link[=read.table]{read.table()}}}
}
\value{
\code{SpatRaster}
}
\description{
Imports a text file into a raster.
}
\details{
A thin wrapper around \link[utils:read.table]{utils::read.table} that creates \code{SpatRaster} objects.
For less exotic file formats, you can use the regular \code{\link[terra:rast]{terra::rast()}} methods.
}
\examples{
# use a system.file to make it work here,
# otherwise just point to your file
# these two are smaller (resampled 0.5) versions of their tif counterpart
# because of data volume limitations

system.file("extdata/l1.txt", package = "mucha") \%>\%
import_txt() \%>\% p()

system.file("extdata/l2.txt", package = "mucha") \%>\%
import_txt() \%>\% p()

}
