% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{kernel}
\alias{kernel}
\alias{kernel_square}
\alias{kernel_circle}
\alias{kernel_gaussian}
\alias{kernel_plot}
\title{Kernel helper functions}
\usage{
kernel_square(size = 3)

kernel_circle(size = 3)

kernel_gaussian(size = 3, sigma = 1)

kernel_plot(x)
}
\arguments{
\item{size}{integer size of the kernel matrix to generate. Must be an odd number so that
the central pixel is actually central.}

\item{sigma}{numeric the sigma parameter for the gaussian kernel}

\item{x}{a kernel to visualize}
}
\value{
a kernel matrix (or a plot for kernel_plot)
}
\description{
A collection of helpers to generate kernel matrices
}
\examples{
kernel_square(3)
kernel_circle(7)
kg <- kernel_gaussian(9, sigma=3)
kg
kg \%>\% kernel_plot()
}
