% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHM.R
\name{MHM}
\alias{MHM}
\title{Multiscale Heterogenity Map}
\usage{
MHM(
  x,
  window = window_quick(),
  kernel = kernel_circle,
  fun = mean,
  na.policy = "omit",
  fillvalue = NA,
  expand = FALSE,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{SpatRaster}

\item{window}{numeric sequence of window sizes (default to \code{c(3, 13, 23)} via \link{window_quick}). Passed to \code{kernel} function}

\item{kernel}{function among \link{kernel}s (default to \code{kernel_circle})}

\item{fun}{function that takes multiple numbers, and returns a numeric vector (one or multiple numbers). For example mean, modal, min or max}

\item{na.policy}{character. Can be used to determine the cells of \code{x} for which focal values should be computed. Must be one of "all" (compute for all cells), "only" (only for cells that are \code{NA}) or "omit" (skip cells that are \code{NA}). Note that the value of this argument does not affect which cells around each focal cell are included in the computations (use \code{na.rm=TRUE} to ignore cells that are \code{NA} for that)}

\item{fillvalue}{numeric. The value of the cells in the virtual rows and columns outside of the raster}

\item{expand}{logical. If \code{TRUE} The value of the cells in the virtual rows and columns outside of the raster are set to be the same as the value on the border. Only available for "build-in" \code{fun}s such as mean, sum, min and max}

\item{na.rm}{logical passed to \code{fun}. Whether to remove NA in the calculation for each focal cell.
Not the NA in the global SpatRaster. See \link[terra:focal]{terra::focal}}

\item{...}{additional arguments passed to \code{fun} such as \code{na.rm}}
}
\value{
\code{SpatRaster}
}
\description{
Description
}
\examples{
# load terra
library(terra)

# import an example file
# and turn it into a SpatRaster
landscape <- import_example("l1.tif")

# plot it
plot(landscape)

# make it smaller for example purpose:
l0 <- landscape \%>\% raster_resample(0.1)

# plot the little landscape
p(l0)

# calculate MHM on it
l0_mhm <- MHM(l0,
              window=seq(2, 13, by=2)+1,
              fun=richness)

# display all intermediate maps
p(l0_mhm)

# the profile plot
ms_profile(l0_mhm)

# and a synthetic map
app(l0_mhm,  median) \%>\% p()
app(l0_mhm,  sd) \%>\% p()

# change the color palette and print side by side
l0_sd <- app(l0_mhm,  sd)
c(l0, l0_sd) \%>\% p()

}
\references{
Gaucherel, C. (2007) Multiscale heterogeneity map and associated scaling profile
for landscape analysis, \emph{Landscape and Urban Planning} 82(3) 95-102.
doi: 10.1016/j.landurbplan.2007.01.022
}
