% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.R
\name{ms_profile}
\alias{ms_profile}
\alias{ms_profile_df}
\title{Profile method for MHM and CMP}
\usage{
ms_profile_df(x)

ms_profile(
  x,
  summary_fun = mean,
  error_fun = se,
  title = "multiscale profile",
  ylab = "mean +/- SE"
)
}
\arguments{
\item{x}{SpatRast typically obtained from \link{MHM}/\link{CMP}}

\item{summary_fun}{a function to calculate a
summary statistic for each window (default to \code{mean})}

\item{error_fun}{a function to calculate an
error statistic for each window (default to standard error)}

\item{title}{for the plot}

\item{ylab}{for the y axis}
}
\value{
a plot or a \code{data.frame}
}
\description{
Profile the result of \link{MHM}/\link{CMP}, ie obtain, summarize and visualize
results for each window size.
}
\examples{
landscape <- import_example("l1.tif") \%>\%
   raster_resample(0.2)

# vanilla MHM
res <- MHM(landscape, window=c(3, 5, 7, 9, 13, 17, 33))

# plot the ms_profile plot directly
res \%>\% ms_profile()

# you can tweak summary_fun, error_fun and title
res \%>\% ms_profile(summary_fun=max, error_fun=var, ylab="max +/- var", title="ms")

# you can also get the raw data.frame
# to make your own (eg ggplot2) graph
res \%>\% ms_profile_df() \%>\% head()

}
