% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{odd_floor}
\alias{odd_floor}
\title{Flooring to the closest odd number}
\usage{
odd_floor(x)
}
\arguments{
\item{x}{numeric}
}
\value{
an odd integer, the closest from the numeric passed
}
\description{
To ease the use in \code{window} size in \link{MHM}/\link{CMP}
}
\examples{
odd_floor(4)
odd_floor(99)
odd_floor(c(54, 53.99, 54.01))
}
