% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_helpers.R
\name{raster_helpers}
\alias{raster_helpers}
\alias{raster_classes}
\alias{raster_nclasses}
\alias{raster_likely_categorical}
\alias{raster_range}
\alias{raster_ntot}
\alias{raster_nNA}
\alias{raster_anyNA}
\alias{raster_declareNA}
\alias{raster_summary}
\alias{raster_resample}
\alias{raster_rescale}
\title{A collection of raster helpers to ease work with landscapes}
\usage{
raster_classes(x)

raster_nclasses(x)

raster_likely_categorical(x)

raster_range(x)

raster_ntot(x)

raster_nNA(x)

raster_anyNA(x)

raster_declareNA(x, NA_value)

raster_summary(x)

raster_resample(x, width, height, method = "near")

raster_rescale(x, new_min = 0, new_max = 255)
}
\arguments{
\item{x}{\link[terra:SpatRaster-class]{terra::SpatRaster}}

\item{NA_value}{which value, typically an integer (either 0 or 255)
should be considered as an NA instead}

\item{width}{numeric either an integer > 1 (new width) or a proportion to downsize width}

\item{height}{numeric either an integer > 1 (new height) or a proportion to downsize height}

\item{method}{default to \code{near}, passed to \link[terra:resample]{terra::resample}}

\item{new_min}{new min value after rescaling (default to 0)}

\item{new_max}{new max value after rescaling (default to 255)}
}
\value{
various classes, depending on the method
}
\description{
Besides those, you could also need
\link[terra:is.bool]{terra::as.factor} or \link[terra:inplace]{terra::set.cats} or \link[terra:classify]{terra::classify}
to make landscape explicitely qualitative.
}
\examples{
library(terra)

# let's continue with other raster_helpers

# import an example file
landscape_quali <- simulate_voronoi(2329)
landscape_quant <- simulate_continuous(2329)

# minimalist and non absolute deduction based on the number of classes
# if <20, likely categorical though
raster_likely_categorical(landscape_quali)
raster_likely_categorical(landscape_quant)

# plot
landscape_quali \%>\% p()
landscape_quant \%>\% p()

# show the number of classes (categorical)
landscape_quali \%>\% raster_classes()

# count the number of classes (categorical)
landscape_quali \%>\% raster_nclasses()

# show raster range (continuous)
landscape_quant \%>\% raster_range()

# a packed summary
raster_summary(landscape_quant)
raster_summary(landscape_quali)

# to decrease size
# this is a good way to explore fast the big trends,
# then run the full map using MHM/CMP

\dontrun{
# this tends to crash on my (antique) machine
# bilinear is likely better for continuous
raster_resample(landscape_quant, width=0.1, method="bilinear")

# bilinear is likely better for continuous
raster_resample(landscape_quali, width=0.1, method="near")
}

}
