% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{se}
\alias{se}
\title{Standard error}
\usage{
se(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
Brings a standard error \code{(var(x)/sqrt(n))} function
}
\details{
SE is the historical error measurement in C. Gaucherel's papers.
}
\examples{
x <- c(1, 2, 3, 4, 3, 4, 3, 2, 1, 4)
sd(x) # standard deviation
se(x) # standard error
}
