% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\alias{simulate_continuous}
\alias{simulate_voronoi}
\title{Simulate categorical and contious landscapes}
\usage{
simulate_continuous(seed = sample(1:1000, 1), nrow = 200, ncol = 200)

simulate_voronoi(
  seed = sample(1:1000, 1),
  n_cat = 6,
  n_points = 42,
  nrow = 200,
  ncol = 200
)
}
\arguments{
\item{seed}{a random seed}

\item{nrow}{for the simulated landscape}

\item{ncol}{for the simulated landscape}

\item{n_cat}{number of patches categories}

\item{n_points}{number of Voronoi cells}
}
\description{
These functions simulate simple categorical and continuous landscapes
using Voronoi and random noise function built in \code{terra}.
}
\examples{
# for the sake of replicability we set seeds

# categorical landscapes
simulate_voronoi(seed=2329) \%>\% p()
simulate_voronoi(seed=1517, nrow = 400, ncol=600, n_cat = 4, n_points=820) \%>\% p()

# continuous landscapes (dem-like)
simulate_continuous(2329) \%>\% p()
simulate_continuous(1517, nrow = 400, ncol=600) \%>\% p()
}
