% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{window}
\alias{window}
\alias{window_quick}
\alias{window_broad}
\alias{window_linearpixels}
\title{Window helper functions}
\usage{
window_quick()

window_broad(x, steps = 3)

window_linearpixels(x, steps = 3)
}
\arguments{
\item{x}{a raster}

\item{steps}{the number of window sizes to generate}
}
\value{
a sequence on integer to pass to \link{MHM}/\link{CMP}
}
\description{
A collection of helpers to generate window sequences matrices
}
\details{
\code{window_quick} has a function shape but simply returns \code{c(3, 13, 23)}.
\code{window_broad} creates a sequence from \code{3} to \code{smallest_dim/3}.
\code{window_linearpixels} tries to have a roughly regular sequence of pixel numbers
which increase with a power two of kernel size otherwise.
}
\examples{
kernel_square(3)
kernel_circle(7)
kg <- kernel_gaussian(9, sigma=3)
kg
kg \%>\% kernel_plot()
}
