% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMP.R
\name{CMP}
\alias{CMP}
\title{Comparison Map Profile}
\usage{
CMP(
  x,
  y,
  window = window_quick(),
  kernel = kernel_circle,
  fun = mean,
  fillvalue = NA,
  ...
)
}
\arguments{
\item{x, y}{rasters as \code{\link[terra:SpatRaster-class]{terra::SpatRaster()}}}

\item{window}{numeric sequence of window sizes (default to \code{c(3, 13, 23)} via \link{window_quick}). Passed to \code{kernel} function}

\item{kernel}{function among \link{kernel}s (default to \code{kernel_circle})}

\item{fun}{function that takes multiple numbers and return a single value such
as \code{mean} (default), \code{median}, \code{min}, \code{max}, terra::modal, \link{richness}, \link{simpson},
\link{shannon}, \link{shannon_evenness} or even custom functions (see vignette)}

\item{fillvalue}{numeric. The value of the cells in the virtual rows and columns outside of the raster}

\item{...}{additional parameters to pass to \link[terra:focalPairs]{terra::focalPairs}}
}
\value{
\code{SpatRaster}
}
\description{
Description
}
\examples{
library(terra)

# import the first example file
l1 <- import_example("l1.tif")

# and the second
l2 <- import_example("l2.tif")

# we can plot them side by side
c(l1, l2) \%>\% p()

# make them smaller by a factor 10 for example purpose
ls1 <- raster_resample(l1, 0.1)
ls2 <- raster_resample(l2, 0.1)

ls_cmp <- CMP(ls1, ls2,
              window=seq(3, 13, by=2),
              fun=diff_rmse)

p(ls_cmp)

ms_profile(ls_cmp)

app(ls_cmp, mean) \%>\% p()

}
\references{
Gaucherel, C., Alleaume, S. and Hély, C. (2008) The Comparison Map Profile method:
a strategy for multiscale comparison of quantitative and qualitative images.
\emph{IEEE Transactions on Geoscience and Remote Sensing} 46 (9): 2708-2719
doi: 10.1109/TGRS.2008.919379
}
