% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhm_funs.R
\name{mhm_funs}
\alias{mhm_funs}
\alias{richness}
\alias{simpson}
\alias{shannon}
\alias{shannon_evenness}
\alias{kappa_index}
\alias{contagion}
\alias{contagion_cpp_r}
\title{Summary metrics on a single raster}
\usage{
richness(x, ...)

simpson(x, ...)

shannon(x, ...)

shannon_evenness(x, ...)

kappa_index(x, ...)

contagion(x, ...)

contagion_cpp_r(x, ...)
}
\arguments{
\item{x}{a fraction of a \code{SpatRaster} raster}

\item{...}{additional parameters,
used to collect \link[terra:focal]{terra::focal} own passing through \code{...}}
}
\value{
numeric the index of interest
}
\description{
Summary functions that return a scalar when passed with a portion of raster.
Typically passed to \link{MHM}.
}
\details{
\itemize{
\item \code{richness} simply returns the total number of unique classes
(only makes sense on categorical landscapes)
\item \code{simpson} is the Simpson's diversity index
\item \code{shannon} is the Shannon's diversity index
\item \code{shannon_evenness} is the Shannon's evenness index
\item \code{kappa_index} is Kappa index
\item \code{student_p} is the p value from a Student's test
\item \code{contagion} is the contagion index (Riitters, 1996) and
also from MHM paper (only makes sense on categorical landscapes)
}
}
\examples{
set.seed(2329) # for reproducibility
x <- sample(1:3, size=49, replace=TRUE, prob=c(0.1, 0.1, 1))  # unbalanced vector
table(x)
x \%>\% richness()
x \%>\% simpson()
x \%>\% shannon()
x \%>\% shannon_evenness()
x \%>\% kappa_index()
x \%>\% contagion()
}
