% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.R
\name{p}
\alias{p}
\title{Monoscale plot after an MHM/CMP}
\usage{
p(x, palette = "viridis", asp = 1, ncol = NULL, title, ...)
}
\arguments{
\item{x}{a \code{SpatRaster} typically obtained after \link{MHM}/\link{CMP}}

\item{palette}{one of \link[grDevices:palettes]{grDevices::hcl.colors} palette description (default to \code{viridis})}

\item{asp}{target aspect for the layout (default to 1 that is a square)}

\item{ncol}{integer number of columns for the layout, if provided \code{asp} is ignored}

\item{title}{for the plot, if missing use \code{names(x)}}

\item{...}{additional parameters to single \code{plot} call}
}
\description{
For what we have to represent, this is a better plot for \code{SpatRaster}
than base \link{terra}'s plot. The main benefit being
sharing the legend across monoscale maps.
}
\examples{
# first calculate a small and simple MHM
l <- import_example("l1.tif") \%>\%
raster_resample(0.1) \%>\%
MHM(window=c(3, 13, 23, 33), fun=richness)

# then let the fun begins
p(l) # compare with plot(l)
p(l, palette = "RdYlBu", ncol=1)
}
